modules/data_processing/pyproject.toml
[tool.poetry]
name = "data processing"
version = "0.1.0"
description = "Data Processing pipeline for AI Vegan Nutritionist"
authors = ["Ravinder Rai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
opensearch-py = "^2.7.1"
requests-aws4auth = "^1.3.1"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
transformers = "^4.44.2"
torch = "^2.4.1"
sentence-transformers = "^3.1.1"
boto3 = "^1.35.23"
langchain = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
